# Conflicts:
#   public/css/general.css
#   resources/lang/ca/titles.php
#   resources/lang/en/titles.php
#   resources/lang/es/titles.php
#   resources/views/detail/detailAnimes.blade.php
#   resources/views/detail/detailFilms.blade.php
#   resources/views/detail/detailSeries.blade.php
#   resources/views/profile/profile.blade.php
#   routes/web.php

+++ app/Http/Controllers/UserController.php
@@ -9,8 +9,10 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use App\Models\User;
 use App\Models\FavoriteList;
+use App\Models\Image;
 use App\Models\Like;
 use App\Models\Review;
+use Illuminate\Support\Facades\App;
 
 class UserController extends Controller
 {
@@ -108,20 +110,49 @@ class UserController extends Controller
 
     public function profileUpdate(Request $request)
     {
+        $user = Auth::user();
+        $userId = $user->id;
+
         $request->validate([
-            'username' =>'required|min:4|string|max:255',
+            'username' =>'required|min:4|string|max:255|unique:users,nick,'.$userId,
             'language' => 'required',
         ]);
         $user = Auth::user();
         $user->nick = $request['username'];
         $user->lang = $request['language'];
 
+        
         $user->save();
+        
+        $this->changeLanguages($request['language']);
+
         return view('profile.profile')->with('message','Profile Updated');
     }
     public function deleteAccount(){
+
         $user = Auth::user();
-        $user->delete();
+        $id = $user->id;
+            
+        $isset_reviews = Review::where('user_id', $id)->first();
+        $isset_likes = Like::where('user_id', $id)->first();
+        
+        if($isset_likes) {
+            $likes = Like::where('user_id', $id)->get();
+            foreach($likes as $like) {
+                $like->where('user_id', $user->id)->delete();
+            }
+        }
+        
+        if($isset_reviews) {
+            $reviews = Review::where('user_id', $id);
+            foreach($reviews as $review) {
+                $review->where('user_id', $id)->delete();
+            }
+        }
+
+
+        $userToDelete = User::findOrFail($id);
+        $userToDelete->delete();
 
         Auth::logout();
         return redirect()->to('register')->with('signOut', 'Cuenta eliminada!');
@@ -141,4 +172,36 @@ class UserController extends Controller
 
         return view('activity.activity', ['likes' => $likes]);
     }
+    
+    public function changeLanguages($lang)
+    {
+        App::setLocale($lang);
+        session()->put('locale', $lang);
+    }
+
+    public function getUserProfileImg()
+    {
+        $images = Image::get();
+        return view ('profile.profileImg', ['images' => $images]);
+    }
+
+    public function postUserProfileImg($id = null)
+    {
+        if(isset($id) && !is_null($id) && !empty($id)){
+            
+            $authUser = Auth::user();
+
+            $user = User::find($authUser->id);
+            $user->image_id = $id;
+            $user->save();
+
+            return redirect()->to(route('user.profile'))->with('profileImageUpdated', trans('profile.img_updated'));
+            
+        }else{
+            return view('profile.profile');
+        }
+    }
+
+
+
 }
